+++ backend/src/main/java/com/woowacourse/thankoo/alarm/infrastructure/SlackUserMapper.java
@@ -1,8 +1,7 @@
-package com.woowacourse.thankoo.alarm.model;
+package com.woowacourse.thankoo.alarm.infrastructure;
 
 import com.woowacourse.thankoo.alarm.infrastructure.dto.SlackUserResponse;
 import com.woowacourse.thankoo.alarm.infrastructure.dto.SlackUsersResponse;
-import com.woowacourse.thankoo.alarm.infrastructure.SlackClient;
 import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
